+++ src/class/pyasync.rs
@@ -8,9 +8,10 @@
 //! [PEP-0492](https://www.python.org/dev/peps/pep-0492/)
 //!
 
+use crate::callback::IntoPyCallbackOutput;
 use crate::derive_utils::TryFromPyCell;
 use crate::err::PyResult;
-use crate::{ffi, PyClass, PyObject};
+use crate::{ffi, IntoPy, IntoPyPointer, PyClass, PyObject, Python};
 
 /// Python Async/Await support interface.
 ///
@@ -60,33 +61,28 @@ pub trait PyAsyncProtocol<'p>: PyClass {
 
 pub trait PyAsyncAwaitProtocol<'p>: PyAsyncProtocol<'p> {
     type Receiver: TryFromPyCell<'p, Self>;
-    type Success: crate::IntoPy<PyObject>;
-    type Result: Into<PyResult<Self::Success>>;
+    type Result: IntoPyCallbackOutput<PyObject>;
 }
 
 pub trait PyAsyncAiterProtocol<'p>: PyAsyncProtocol<'p> {
     type Receiver: TryFromPyCell<'p, Self>;
-    type Success: crate::IntoPy<PyObject>;
-    type Result: Into<PyResult<Self::Success>>;
+    type Result: IntoPyCallbackOutput<PyObject>;
 }
 
 pub trait PyAsyncAnextProtocol<'p>: PyAsyncProtocol<'p> {
     type Receiver: TryFromPyCell<'p, Self>;
-    type Success: crate::IntoPy<PyObject>;
-    type Result: Into<PyResult<Option<Self::Success>>>;
+    type Result: IntoPyCallbackOutput<IterANextOutput>;
 }
 
 pub trait PyAsyncAenterProtocol<'p>: PyAsyncProtocol<'p> {
-    type Success: crate::IntoPy<PyObject>;
-    type Result: Into<PyResult<Self::Success>>;
+    type Result: IntoPyCallbackOutput<PyObject>;
 }
 
 pub trait PyAsyncAexitProtocol<'p>: PyAsyncProtocol<'p> {
     type ExcType: crate::FromPyObject<'p>;
     type ExcValue: crate::FromPyObject<'p>;
     type Traceback: crate::FromPyObject<'p>;
-    type Success: crate::IntoPy<PyObject>;
-    type Result: Into<PyResult<Self::Success>>;
+    type Result: IntoPyCallbackOutput<PyObject>;
 }
 
 #[doc(hidden)]
@@ -107,37 +103,34 @@ impl ffi::PyAsyncMethods {
     where
         T: for<'p> PyAsyncAnextProtocol<'p>,
     {
-        self.am_anext = anext::am_anext::<T>();
+        self.am_anext = am_anext::<T>();
     }
 }
 
-mod anext {
-    use super::PyAsyncAnextProtocol;
-    use crate::callback::IntoPyCallbackOutput;
-    use crate::err::PyResult;
-    use crate::IntoPyPointer;
-    use crate::Python;
-    use crate::{ffi, IntoPy, PyObject};
+pub struct IterANextOutput(Option<PyObject>);
 
-    struct IterANextOutput<T>(Option<T>);
-
-    impl<T> IntoPyCallbackOutput<*mut ffi::PyObject> for IterANextOutput<T>
-    where
-        T: IntoPy<PyObject>,
-    {
-        fn convert(self, py: Python) -> PyResult<*mut ffi::PyObject> {
-            match self.0 {
-                Some(val) => Ok(val.into_py(py).into_ptr()),
-                None => Err(crate::exceptions::StopAsyncIteration::py_err(())),
-            }
+impl IntoPyCallbackOutput<*mut ffi::PyObject> for IterANextOutput {
+    fn convert(self, _py: Python) -> PyResult<*mut ffi::PyObject> {
+        match self.0 {
+            Some(o) => Ok(o.into_ptr()),
+            None => Err(crate::exceptions::StopAsyncIteration::py_err(())),
         }
     }
+}
 
-    #[inline]
-    pub(super) fn am_anext<T>() -> Option<ffi::unaryfunc>
-    where
-        T: for<'p> PyAsyncAnextProtocol<'p>,
-    {
-        py_unarys_func!(PyAsyncAnextProtocol, T::__anext__, IterANextOutput)
+impl<T> IntoPyCallbackOutput<IterANextOutput> for Option<T>
+where
+    T: IntoPy<PyObject>,
+{
+    fn convert(self, py: Python) -> PyResult<IterANextOutput> {
+        Ok(IterANextOutput(self.map(|o| o.into_py(py))))
     }
+}
+
+#[inline]
+fn am_anext<T>() -> Option<ffi::unaryfunc>
+where
+    T: for<'p> PyAsyncAnextProtocol<'p>,
+{
+    py_unarys_func!(PyAsyncAnextProtocol, T::__anext__)
 }